Added `ser/maps` module
