+++ crates/bevy_reflect/src/serde/ser/serializer.rs
@@ -1,8 +1,9 @@
 use crate::serde::ser::enums::EnumSerializer;
+use crate::serde::ser::maps::MapSerializer;
 use crate::serde::ser::structs::StructSerializer;
 use crate::serde::ser::tuple_structs::TupleStructSerializer;
 use crate::serde::ser::tuples::TupleSerializer;
-use crate::serde::{ArraySerializer, ListSerializer, MapSerializer, Serializable, SetSerializer};
+use crate::serde::{ArraySerializer, ListSerializer, Serializable, SetSerializer};
 use crate::{PartialReflect, ReflectRef, TypeRegistry};
 use serde::ser::{Error, SerializeMap};
 use serde::Serialize;
@@ -164,11 +165,9 @@ impl<'a> Serialize for TypedReflectSerializer<'a> {
                 registry: self.registry,
             }
             .serialize(serializer),
-            ReflectRef::Map(value) => MapSerializer {
-                map: value,
-                registry: self.registry,
+            ReflectRef::Map(value) => {
+                MapSerializer::new(value, self.registry).serialize(serializer)
             }
-            .serialize(serializer),
             ReflectRef::Set(value) => SetSerializer {
                 set: value,
                 registry: self.registry,
